+++ src/standards_atlas/application/formal_semantics/__init__.py
@@ -1,5 +1,6 @@
 """Application support for packaged formal ontologies."""
 
+from .extraction_projector import SemanticExtractionProjectionAugmenter
 from .ontology_definition import FormalOntologyDefinition
 from .projector import DeterministicFormalSemanticProjector
 from .resource_repository import ResourceFormalOntologyRepository
@@ -8,4 +9,5 @@ __all__ = [
     "DeterministicFormalSemanticProjector",
     "FormalOntologyDefinition",
     "ResourceFormalOntologyRepository",
+    "SemanticExtractionProjectionAugmenter",
 ]